+++ CHANGELOG.md
@@ -14,6 +14,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * `dset.view_sql` now has a `memory` kwarg that will force the database to live in memory.
 * The SQLite backend now agrees with the dict backend and does not map None to values for image names and file names.
 
+### Removed
+* `kwcoco.utils.util_slice` use `kwarray.util_slice` instead.
+
 
 ## Version 0.2.0 - Released 2021-04-22